Added code to search for contact details.
The code checks if the drop down menu and the filter text box has values, if it does it then checks them against the database. If a result was found, it is displayed in an alert box. If a result was not found, an error message is displayed in an alert box. If the drop down or the filter text box or both are empty, an error message is displayed in an alert.